split elements
