+++ src/interface/global.ts
